+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_shelf_place_v2.py
@@ -36,10 +36,10 @@ class SawyerShelfPlaceEnvV2(SawyerXYZEnv):
         self.num_resets = 0
 
         self._random_reset_space = Box(
-            np.hstack((obj_low, goal_low)),
-            np.hstack((obj_high, goal_high)),
+            np.hstack((obj_low, goal_low)).astype(np.float32),
+            np.hstack((obj_high, goal_high)).astype(np.float32),
         )
-        self.goal_space = Box(np.array(goal_low), np.array(goal_high))
+        self.goal_space = Box(np.array(goal_low, dtype=np.float32), np.array(goal_high, dtype=np.float32))
 
     @property
     def model_name(self):
